Major rewrite of GlobalUserrights to unbreak it for 1.29
This extension needed attention as it did not function at all after the
changes made to Special:Userrights with 1.29, which this extension
derives of.
This bumps the version number up to 1.5 to prevent confusion with
ShoutWiki's 1.4-SW release.
Things fixed:
* Support for SQLite
* Support for expiring group membership
* Log formatting is now done with LogFormatter instead of the previous
hack, identical to the user rights log, complete with the advantages
that brings
* User Uid's are now the Central ID. This doesn't change anything for
shared tables setups, but should make the extension easier to support on
setups that have a different Central ID provider
* Global group membership has now been abstracted similar to local group
membership
* A hook handler for UsersPagerDoBatchLookups has been added, because a
previous hook is not in MediaWiki (anymore?)
* Short array syntax used

Bug: T166436
Change-Id: I2eca420b3590ce7680cbc96c3432ff777e091d2a